Attempt at removing routes.dart dependency.
By using a separate AnimationController unrelated to the one in the
route. The problem seems to be that the destination route during a back
gesture (the old route) also needs an animation, so that it can animate
itself as it gets peeked on by the top route. It's hard to coordinate
that without using the route/navigator, though.